Refactor packet ordering utilities
